remove creator and edit current projectWizard to custom view
